initial image recognition
For now, the dominoes need to be arranged so the bottom left corner is
down, upper left corner is left, upper right is up, and right bottom is
right.  After taking the picture, press count -> shapes -> return to get
the domino back to the game window.  It can detect multiple dominoes,
but not every well.

The device needs to stay in the same orientation the entire time
otherwise the picture will go away, and the game window needs to be
empty for it to go in, probably.

I don't think openCV is being used anymore.