Improved accuracy, bug fixes
Domino orientation restriction remains, but it is much more accurate at detecting single and multiple dominoes now.

Changed the inputs on the camera page to reflect the following:
sigma - strength of blur
maskSize- size of the mask used when applying blur
limit - distance that can be traveled while edge detecting
checkLimit - how many spaces can be skipped when looking for continuous edges
percent - percent of the total pixels that will be used to determine the high and low thresholds, can be set to -1 to automatically determine percent based on .. stuff

picture is recalculated on orientation change now, I could have used a button, but I felt I needed to sell the platform by making use of the motion controls.