Somewhat Working Scrabble
Scrabble is in a somewhat working state now.
There's a bunch of bugs: like ScrabbleLetter lying about which one it
really is.
The big one though is that the get method to Anagramica doesn't work.
Something's gone wrong, and I'm not sure what.
Still, other than calculation, the app is stable enough to push now.