more detection improvements
fixed entire sides being dropped, was calculating the area of a side
assuming it was a square, which is rarely actually the case.  When the
actual area of a side was larger than area calculated as a square then
all the points would be considered outside the bounds and dropped.