Cleanup, refactoring and optimizations
Renamed DetectedShape to DominoHandler

Renamed DominoDetection to ObjectFinder

Created an interface for DominoHandler and other shape detectors to
inherit from.   At this point MainActivity should be able to be made
into an abstract class or interface as the only domino specific part of
the code is creating a DominoHandler object, and that can be done
elsewhere.

Moved shape categorization out of ObjectFinder and into ObjectHandler.
This is done by using the objectList and cornerList from ObjectFinder in
the ObjectHandler method readObjectList.

Renamed various methods and variable names that were misleading.

Changed peak and edge arrays in ObjectFinder to boolean since they were
0 or 255 anyway.

Added descriptions to the parameters used in canny as well as possible
optimizations.